I could only get the FCMissionReader to compile on Mac.
While I developed this engine to use OpenGLES 2 I already knew
that OpenGLES 2 was deprecated on the Mac OS ecosystem. I did
make this rendering engine on a Raspberry PI 4 after all.

I initially planned to use a wrapper like metalGL or ANGLE.
I soon realized that both of them has their pitfalls.
metalGL is not free software, so I feel like it would not be
right to use it.

Also, ANGLE might not support Mac OS itself in metal, but
it might support OpenGLES to OpenGL layer.
I tried to force using the full OpenGL 2 in order to get the
graphics working, but I did not succeed. However, it only made
the screen black.

So, I might give ANGLE a try on the Mac once I get back to this :)