Now, the x64 linux version has problems.
It turned out that there is legitmatly wrong with the code in SDL2
and OPENGLES 2.

The only successful compilation is on the Raspberry Pi 4 Manjaro
linux with the FCMissionReader and everything that has to do with
the graphics.

I am going to try to diagnose the problem, but for now I cannot
get the graphics to work.

While I was developing, I should have checked if the graphics code
worked on my main computer more constantly.