I managed to find a way to get the graphics working on x64.
I discovered that my initialization code has a huge flaw in it.
I created a SDL_Renderer when it was actually not needed, because
I already had a context.

So what I did is get rid of the renderer. I will check if this
will compile on a Raspberry PI 4.

I also realized that I do not actually need a wrapper to turn
OpenGLES 2 into OpenGL 2 because OpenGLES 2 is a subset of
OpenGL 2. Meaning that standard OpenGLES 2 would work just fine
wine OpenGL 2.

This code might work on the MAC. However, I must add options to
cmake in order to get this to work.