I made yet another "proof" for this program.
This proof isolates the bits inside the opcode bitfield of the
bones that decide which location unit or rotation unit is
constant and non-constant.

If this proof is successful then I could get rid of that table.

This table is flawed, and it makes the program potentially less
accurate.

Right now I am checking if I am wrong on 0x0D since it does not
agree with the proof unless there is a special condition created.