I made improvements to the bone animation.
I removed that table. That table is not only bad pratice, but
it took up unnessary memory.

I replaced with a better system. A struct defining the bitfield
is what was needed.

Now, the code feels a bit more elegant. (Although, I need to still
need to revise my code).