I had modified the TilResource to use the Buffer Reader API.
I realized that the sound is broken, so this is next in my list to
fix.

However, I think I had at least transitioned all the stuff to the
new API.

The next thing to refractor now is the Audio API. I think it
copies the audio data too many times anyways.

WAV should be supported in Utilities rather than mission also.