Updated the project to have more safe guards.
I made the build system use GLEW if it is found. This can be used
for the OpenGL loading system.

I made it so the system would attempt different OpenGL context
core versions. ES is tried first, CORE is tried second, COMPAT is
tried the last.

The primary problem (of the many problems) with the Windows build
is that SDL is failing to force OpenGL 2.