Interesting Bug
This has a bug where the texture system is broken. I simply wanted to
know what it looked like.

This commit is big.

First, PNG loading and reading is disabled for now.

I made sure that each resource that is capable of exporting a texture
to use the new image input and output system. This makes it so
using more formats are now possiable.

I enabled the qoi exporter to be able to export black and white images
by converting the black and white to colors.

I also changed TilResource and ObjResource to read the textures after
they had been parsed, so they would know their file names.

I also changed the ModelBuilder to have less hardcoded behavior with
PNG.