I fixed a bug in the Triangle class.
I have fixed a bug where the triangle does not fully update its
calculated plane when changed by setPoint(). I seen this bug while
developing the tester class. In a matter of fact, I fixed the bug
before I wrote code to test it. I checked the tester code to see
if it will catch the error and it did.