This is the new image system.
* Grid2D is an abstract class that would contain a grid system.
* GridReg2D stores the slots of the grid in a normal grid to row
scheme.
* GridSq2D stores slots that is optimized for the CPU cache. One
limitation of this algorithm is that the height and width to be
power by two, and for the grid to have the same width and height.
* PixelFormat holds color information representations. It can
handle converting between different color formats. I finally
added 15 bit color support.
* Image2D combines all the improvements of the PixelFormat and
the GridReg2D, and makes an image.

I need to make automated tests before using all these new classes.