The PixelFormat header and source are revised.
I do not feel like making a BIT aligned ImageFormat.

Instead what I feel like doing is to only support reading and
writing on color paletted images. The paletted images would only
store each index in 1 unsigned byte.

This would simplify the code.

Making a new implementation of the image format proved to be quite
the time sink. However, I feel like it would be worth it. It will
improve the readablity of the code.

In addition, it would make the software renderer easier to write.