I found out that Cobj's normals are optional.
Belive it or not some kind of bounds checking excluded cobj's.
It is not a face reading issue like I had initially belived.

Right now my opengl implementation crashes when using the normaless models.

The workaround would be to simply make a dummy normal value, and
apply it to the entire model.