This now system is ready to use.
I think Environment is the only thing with the old system.

I have not forgotten the Windows port. I plan to write an
alturnative rendering system in the more distant future.
I will write a software renderer to render text only for now.
It will just be enough for Windows.