I think I made my preperations.
Now that the logicals had been placed in it is time to do this
vague algorithm.

So, I am just going to write down an algorithm with an assumption.
The top left corner would immediately inherit the color that the
graphics tile has. The top right corner would read from quadrant
opposite from the x axis of the quadrant to the current tile. The
bottom left would read the y axis counter part quadrant. The
bottom right would read from the exact opposite quadrant.