A significant upgrade to the font has been made.
There are now effectively two font files.
The playstation font file has a resource id of 1. It can represent
playstation symbols and be small. The lowercase letters are
represented in uppercase letters to save space.

The Windows font file has a resource id of 2. The font is twice as
big, and it includes lower case letters.

The original internal font file has been replaced, but its primary
decendent is the playstation font.

All three files are created by Ghoster738.