Revised the reading
I realized that I might have been reading this font wrong this whole
time. However, now that I made the unit test I could try reading it this
way.

It makes more sense that '@' marks the begining of the font.

Also the number 0x100 is probably the width of the texture.