GJK Collision Detection is being implemented.
I am finally implementing this collsion for the graphics context.

I want to implement transparancy, but I realized the potential frame
rate issue with sorting every polygon in the map.

So, I am implementing occulusion culling for the World.

Which would then be used for a basis culling other objects inside the
graphics context.