GJK Speed OPtimizations.
The Raspberry PI 4 seems to not like my culling method.

Case and Point a single core usage on the SBC is about 80% when culling
is enabled. I need to find a way to make this faster.

The benchmark/stress test seems to indicate that this GJK algorithm is
not slow.