Gradient vertex animation improvements.
The animations for the gradient vertex animations are wrong.

The way to fix this is to first change the direction of the vertex
gadient. However, this would cause an issue where the line is not
connected due to the way the quads are turned into triangles. To solve
this issue the quads would have to be broken up differently