I found out what caused the flickering bug.
I think what happens is that OpenGLES 2 of this specific
device did not account for texture uploading.

It takes time for the textures to get ready, and the VideoCore
drivers did not make the drawing routines wait.

This new approach uploads all the textures first.
Then it draws the mesh information of the map.