Attempted to optimize dynamic triangle data.
There is a rendering bug that looks cool.

I realized what cased this bug, and already have a fix for it.

Dynamic triangles renderer is going to have a more customizable
finalize.

glFinish() is a performance killer for the mac platform. By default,
the bitfield is going to be empty. It should only be enabled in case of
graphical glitches.